fixed two bugs
assertArrayEquals instead of assertEquals
authUser.getId().longValue() == userId.longValue()

+++ Source/uTutor/src/main/java/ch/ututor/controller/ProfileViewController.java
@@ -31,7 +31,7 @@ public class ProfileViewController {
     	User user = null;
     	User authUser = authenticatedUserLoaderService.getAuthenticatedUser();
     	
-    	if( userId == null || authUser.getId() == userId ) {
+    	if( userId == null || authUser.getId().longValue() == userId.longValue() ) {
     		model.addObject( "ownProfile", true );
     		user = authUser;
     	} else {
